data id added
the emitter sends the data with a data id (did). If ack fails to reach the emitter, the same did will be used. The gateway can then determine if the data is new or duplicated. this will prevent the duplication of data in the database.

Two different consecutive data cannot use the same did.