time update separated from influxdb
Previously, time was configured via wifi in the InfluxDbServerConnect function. Now, a new private library with timeConfigWifi is in charge of updating local time, just after configuring wifi and before connectin to InfluxDbServer.

Time configuration can be found in timeCfg.h

A time config timer is implemented. Right now, a large one (20s) is used in order to match the timeout used by the influxDb library.