getGpsPosition, gpsConfig functions, gpsCfg.h file
gpsConfig() sets the baud rate for the UART connection with the GPS module.

getGpsConnection returns latitude and longitude if there are new gps messages, and if they are valid. Otherwise it returns error codes.

New file gpsCfg.h file added for the Gps configuration. For now, it only affects the UART conection.

New library added "EspSoftwareSerial" for conectiong with the GPS module.

New objects gps and ss, from classes defined in the included libraries.