new_value (fullnes) is 8 bytes
in_packet reduced by one byte, since fullness is receives as a percentage.

New functios in collection.cpp to get the file in bins information, and to update the fullness of each bin. In case a bin has been already updated, an error code is returned. The same if the emmitter_id does not match any bin id.