gateway interacts with truck driver
with COL_DONE_BTN driver can say that the current cluster has been collected. Then, gateway looks at the bins table. If some bin in the next cluster is not updated, it tells the driver to follow the route. When info is gathered, if there are full bins, the gateway says to the driver which bins are to be collected. If not, then the next cluster is checked. When all bins are checked empty or emptied, then the gateway says that "all bins have been collected".